# quarkus-astro-app/src/main/webui/api/openapi.yaml
...
openapi: 3.1.0
...
components:
  schemas:
...
    CompletionHistoryDTO:
      type: object
      properties:
        date:
          $ref: "#/components/schemas/LocalDate"
        count:
          type: integer
          format: int64
    CompletionStatsDTO:
      type: object
      properties:
        todayCount:
          type: integer
          format: int64
        weekCount:
          type: integer
          format: int64
        totalCount:
          type: integer
          format: int64
        completionRate:
          type: number
          format: double
    LocalDate:
      type: string
      format: date
      examples:
      - 2022-03-10
...
paths:
...
  /api/stats/history:
    get:
      summary: Get completion history
      description: Returns daily task completion counts for the specified number of
        days
      tags:
      - Statistics
      parameters:
      - description: "Number of days to include in history (1-365, default 30)"
        schema:
          maximum: 365
          minimum: 1
          type: integer
          default: 30
        name: days
        in: query
      responses:
        "200":
          description: History retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CompletionHistoryDTO"
        "400":
          description: Invalid days parameter
        "401":
          description: Not authenticated
  /api/stats/summary:
    get:
      summary: Get completion statistics
      description: "Returns task completion statistics including today, week, total\
        \ counts and completion rate"
      tags:
      - Statistics
      responses:
        "200":
          description: Statistics retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompletionStatsDTO"
        "401":
          description: Not authenticated
...
info:
  title: Task Manager API
  version: 1.0.0
  description: RESTful API for Task Manager application demonstrating Islands Architecture
  contact:
    name: Task Manager Team
servers:
- url: http://localhost:7171
  description: Auto generated value
- url: http://0.0.0.0:7171
  description: Auto generated value