# quarkus-astro-app/src/main/webui/src/lib/api/model/localDate.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Task Manager API
 * RESTful API for Task Manager application demonstrating Islands Architecture
 * OpenAPI spec version: 1.0.0
 */

export type LocalDate = string;